requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.2
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
python-dotenv==0.21.1
requests==2.31.0
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.4
zipp==3.15.0
